backend/requirements.txt
# Core dependencies for the coupon filtering project
requests>=2.31.0
beautifulsoup4>=4.12.2
Pillow>=10.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# FastAPI dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Google API dependencies
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.88.0

# Google Cloud dependencies for real-time notifications
google-cloud-pubsub>=2.20.1

# Authentication and Database
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
sqlalchemy>=2.0.43
alembic>=1.16.4
psycopg2-binary>=2.9.0
